added separate files for the model, view, and controller.
