+++ text_editor.py
@@ -254,55 +254,3 @@ class WindowedLines:
                 self.input_phoneme.update_phonemes(chr(key_input))
         else:
             self.insert(chr(key_input))
-
-class View:
-    def __init__(self, window:curses.window):
-        self.window = window
-        self.phoneme_panel:curses.window = None
-
-    def toggle_panel(self, model:WindowedLines) -> curses.window:
-        if model.get_phoneme_mode():
-            self.phoneme_panel = self.window.subwin(self.window.getmaxyx()[0]-5, 0)
-            self.phoneme_panel.refresh()
-        else:
-            self.phoneme_panel = None
-        return self.phoneme_panel
-
-    def add_str_to_window(self, text:str):
-        self.window.addstr(text)
-        self.window.refresh()
-
-    def update_panel(self, text:str):
-        if self.phoneme_panel:
-            self.phoneme_panel.addstr(text)
-            self.phoneme_panel.refresh()
-
-    def update(self, model:WindowedLines):
-        self.window.erase()
-        self.window.addstr(model.print_window())
-        self.window.move(len(model.prev_lines)-model.top_window_row,min(model.cursor_position, model.window_size[1]))
-        self.window.refresh()
-        self.update_panel(text=model.get_panel_text())
-        self.toggle_panel(model=model)
-
-class Controller:
-    """The connection between the model and the view"""
-    def __init__(self, model:WindowedLines, view:View, window:curses.window):
-        self.model = model
-        self.view = view
-        self.window = window
-
-    def run(self):
-        "the loop connecting the model to user input, displayed using a curses view."
-        curses.noecho()
-        curses.cbreak()
-        curses.raw()
-
-        while self.model.is_running():
-            self.model.putch(self.window.getch())
-            self.view.update(model=self.model)
-
-        curses.nocbreak()
-        curses.echo()
-        curses.endwin()
-
